+++ plone/namedfile/tests/test_blobfile.py
@@ -22,7 +22,7 @@ from plone.namedfile.file import NamedBlobImage
 from plone.namedfile.interfaces import INamedBlobFile
 from plone.namedfile.interfaces import INamedBlobImage
 from plone.namedfile.interfaces import IStorage
-from plone.namedfile.tests.base import NamedFileLayer
+from plone.namedfile.testing import PLONE_NAMEDFILE_INTEGRATION_TESTING
 from plone.namedfile.tests.test_image import zptlogo
 from zope.component import provideUtility
 from zope.interface.verify import verifyClass
@@ -33,15 +33,31 @@ import unittest
 
 
 def registerUtilities():
-    provideUtility(storages.StringStorable(), IStorage, name="__builtin__.str")  # noqa
-    provideUtility(storages.UnicodeStorable(), IStorage, name="__builtin__.unicode")  # noqa
-    provideUtility(storages.FileChunkStorable(), IStorage, name="plone.namedfile.file.FileChunk")  # noqa
-    provideUtility(storages.FileDescriptorStorable(), IStorage, name="__builtin__.file")  # noqa
+    provideUtility(
+        storages.StringStorable(),
+        IStorage,
+        name="__builtin__.str"
+    )
+    provideUtility(
+        storages.UnicodeStorable(),
+        IStorage,
+        name="__builtin__.unicode"
+    )
+    provideUtility(
+        storages.FileChunkStorable(),
+        IStorage,
+        name="plone.namedfile.file.FileChunk"
+    )
+    provideUtility(
+        storages.FileDescriptorStorable(),
+        IStorage,
+        name="__builtin__.file"
+    )
 
 
 class TestImage(unittest.TestCase):
 
-    layer = NamedFileLayer
+    layer = PLONE_NAMEDFILE_INTEGRATION_TESTING
 
     def setUp(self):
         registerUtilities()